:construction::truck: nav mixins into mgnifylist/detail
